Rename folder and files to plural Cards
