+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/richardtyler/turing/3mod/projects/rancid/src/index.js":"1","/Users/richardtyler/turing/3mod/projects/rancid/src/reportWebVitals.js":"2","/Users/richardtyler/turing/3mod/projects/rancid/src/components/App/App.js":"3","/Users/richardtyler/turing/3mod/projects/rancid/src/components/Movies/Movies.js":"4","/Users/richardtyler/turing/3mod/projects/rancid/src/components/Card/Card.js":"5","/Users/richardtyler/turing/3mod/projects/rancid/src/components/Header/Header.js":"6","/Users/richardtyler/turing/3mod/projects/rancid/src/components/Film/Film.js":"7","/Users/richardtyler/turing/3mod/projects/rancid/src/components/App/movie-data.js":"8"},{"size":515,"mtime":1612474498568,"results":"9","hashOfConfig":"10"},{"size":362,"mtime":1612305141810,"results":"11","hashOfConfig":"10"},{"size":1244,"mtime":1612475927712,"results":"12","hashOfConfig":"10"},{"size":429,"mtime":1612474498560,"results":"13","hashOfConfig":"10"},{"size":674,"mtime":1612475549846,"results":"14","hashOfConfig":"10"},{"size":339,"mtime":1612474498545,"results":"15","hashOfConfig":"10"},{"size":3142,"mtime":1612497788416,"results":"16","hashOfConfig":"10"},{"size":12967,"mtime":1612474498536,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"19nnvbj",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"25","usedDeprecatedRules":"20"},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"/Users/richardtyler/turing/3mod/projects/rancid/src/index.js",[],["36","37"],"/Users/richardtyler/turing/3mod/projects/rancid/src/reportWebVitals.js",[],"/Users/richardtyler/turing/3mod/projects/rancid/src/components/App/App.js",["38"],"import React, { Component } from 'react';\nimport Header from '../Header/Header';\nimport Movies from '../Movies/Movies';\nimport Film from '../Film/Film';\nimport movieData from './movie-data';\nimport './App.css';\n\nclass App extends Component {\n  constructor() {\n    super();\n    this.state = {\n      movies: movieData.movies,\n      currentMovie: '',\n      // {\"movie\": {id: 1, title: \"Fake Movie Title\", poster_path: \"https://image.tmdb.org/t/p/original//7G2VvG1lU8q758uOqU6z2Ds0qpA.jpg\", backdrop_path: \"https://image.tmdb.org/t/p/original//oazPqs1z78LcIOFslbKtJLGlueo.jpg\", release_date: \"2019-12-04\", overview: \"Some overview that is full of buzzwords to attempt to entice you to watch this movie! Explosions! Drama! True love! Robots! A cute dog!\", average_rating: 6, genres: [{id: 18, name:\"Drama\"}], budget:63000000, revenue:100853753, runtime:139, tagline: \"It's a movie!\" }}\n    }\n  }\n\n  selectMovie = event => {\n    this.setState({ currentMovie: event.target.id });\n  }\n\n  render() {\n    return (\n      <>\n        <Header />\n        <div className=\"App\">\n          {/* {this.state.currentMovie && <Film />}\n          <Movies movies={this.state.movies} /> */}\n          <Film />\n        </div>\n      </>\n    );\n  }\n}\n\nexport default App;\n","/Users/richardtyler/turing/3mod/projects/rancid/src/components/Movies/Movies.js",[],"/Users/richardtyler/turing/3mod/projects/rancid/src/components/Card/Card.js",[],"/Users/richardtyler/turing/3mod/projects/rancid/src/components/Header/Header.js",[],"/Users/richardtyler/turing/3mod/projects/rancid/src/components/Film/Film.js",[],"/Users/richardtyler/turing/3mod/projects/rancid/src/components/App/movie-data.js",[],{"ruleId":"39","replacedBy":"40"},{"ruleId":"41","replacedBy":"42"},{"ruleId":"43","severity":1,"message":"44","line":3,"column":8,"nodeType":"45","messageId":"46","endLine":3,"endColumn":14},"no-native-reassign",["47"],"no-negated-in-lhs",["48"],"no-unused-vars","'Movies' is defined but never used.","Identifier","unusedVar","no-global-assign","no-unsafe-negation"]
+[{"/Users/waffle/Desktop/turing/mod3/projects/rancid/src/index.js":"1","/Users/waffle/Desktop/turing/mod3/projects/rancid/src/reportWebVitals.js":"2","/Users/waffle/Desktop/turing/mod3/projects/rancid/src/components/App/App.js":"3","/Users/waffle/Desktop/turing/mod3/projects/rancid/src/components/App/movie-data.js":"4","/Users/waffle/Desktop/turing/mod3/projects/rancid/src/components/Movies/Movies.js":"5","/Users/waffle/Desktop/turing/mod3/projects/rancid/src/components/Header/Header.js":"6","/Users/waffle/Desktop/turing/mod3/projects/rancid/src/components/Film/Film.js":"7","/Users/waffle/Desktop/turing/mod3/projects/rancid/src/components/Cards/Cards.js":"8"},{"size":515,"mtime":1612532800860,"results":"9","hashOfConfig":"10"},{"size":362,"mtime":1612304232210,"results":"11","hashOfConfig":"10"},{"size":897,"mtime":1612572054144,"results":"12","hashOfConfig":"10"},{"size":12967,"mtime":1612532800852,"results":"13","hashOfConfig":"10"},{"size":478,"mtime":1612573065408,"results":"14","hashOfConfig":"10"},{"size":401,"mtime":1612571872571,"results":"15","hashOfConfig":"10"},{"size":2642,"mtime":1612572840776,"results":"16","hashOfConfig":"10"},{"size":356,"mtime":1612573093059,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"1qu9285",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/waffle/Desktop/turing/mod3/projects/rancid/src/index.js",[],["35","36"],"/Users/waffle/Desktop/turing/mod3/projects/rancid/src/reportWebVitals.js",[],"/Users/waffle/Desktop/turing/mod3/projects/rancid/src/components/App/App.js",[],"/Users/waffle/Desktop/turing/mod3/projects/rancid/src/components/App/movie-data.js",[],"/Users/waffle/Desktop/turing/mod3/projects/rancid/src/components/Movies/Movies.js",[],"/Users/waffle/Desktop/turing/mod3/projects/rancid/src/components/Header/Header.js",[],"/Users/waffle/Desktop/turing/mod3/projects/rancid/src/components/Film/Film.js",[],"/Users/waffle/Desktop/turing/mod3/projects/rancid/src/components/Cards/Cards.js",[],{"ruleId":"37","replacedBy":"38"},{"ruleId":"39","replacedBy":"40"},"no-native-reassign",["41"],"no-negated-in-lhs",["42"],"no-global-assign","no-unsafe-negation"]
